+++ mce.ini
@@ -13,7 +13,7 @@ ModulePath=/usr/lib/mce/modules
 #
 # List of base modules to load
 # Note: the name should not include the "lib"-prefix
-Modules=lock-tklock;lock-devlock;power-dsme;display-dev;x11-ctrl;inactivity-dev;battery-upower;inactivity-inhibit;filter-brightness-als-iio;alarm
+Modules=lock-tklock;lock-devlock;power-dsme;display;x11-ctrl;inactivity;battery-upower;inactivity-inhibit;filter-brightness-als-iio;alarm
 
 # Additional Modules the user wants to load
 # Copy this key to mce.ini.d/99-user.ini and edit it there
